requirements.txt
fastapi
uvicorn
openai>=1.0.0
python-dotenv
pandas
openpyxl
numpy
langchain
langchain-community
langchain-openai
faiss-cpu
nltk
gunicorn
rapidfuzz
